** SECURITY FIX **
Important security fix.
Essentials Plus Plus hooks into existing Essentials functions such as AFK and Warps. The issue that wasn't apparent at first is that the permissions to use the commands happen after the over-ridden code. So players will be able to /setwarp /delwarp /afk without proper permissions. The permissions have been fixed so that permissions are checked prior to executing the hooked in code.

-Also added /uberslap <player> for God_Doesnt_Exist lol